move FeedInspector to its own file
